requirements.txt
PyQt6
PyExifTool
opencv-python
Pillow
